+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.3 - 2020-08-20
+### Added
+- You can now output all available video thumbnails
+- You can now output all available preview thumbnails
+
 ## 1.0.2 - 2020-08-19
 ### Changed
 - Some basic descriptions about the plugin